# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkoutWatch App OpenApi API Documentation",
        "description": "SkoutWatch App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://skoutwatch.test",
            "description": "Local API server"
        },
        {
            "url": "https://dev-api.skoutwatch.com",
            "description": "Staging API server"
        },
        {
            "url": "https://staging.skoutwatch.com",
            "description": "Staging API server"
        },
        {
            "url": "https://api.skoutwatch.com",
            "description": "Live API server"
        }
    ],
    "paths": {
...
        "/api/v1/user/document/resend/otp": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Resend of a registered user",
                "description": "Resend of a registered user",
                "operationId": "userResendDocumentOTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDocumentResendOtpFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign Up a new user",
                "description": "Returns a newly registered user data",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signup",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign In a registered user",
                "description": "Returns a newly registered user data",
                "operationId": "signIn",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/update": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Profile of a registered user",
                "description": "Profile of a registered user",
                "operationId": "updateUserProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout a registered user",
                "description": "Logout a registered user",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Profile of a registered user",
                "description": "Profile of a registered user",
                "operationId": "userProfile",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/dashboard": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Dashboard of a registered user",
                "description": "Dashboard of a registered user",
                "operationId": "userDashboard",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthUpdateFormRequest": {
                "title": "User Update Form Request Fields",
                "description": "User Update request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "first_name": {
                        "title": "First Name",
                        "description": "first name of the user",
                        "type": "string",
                        "example": "Bronx"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last name of the user",
                        "type": "string",
                        "example": "Bronx"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "phone of the user",
                        "type": "string",
                        "example": "+23400233300"
                    },
                    "gender": {
                        "title": "gender",
                        "description": "gender of the user",
                        "type": "string",
                        "example": "m/f/o"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Address of the user",
                        "type": "string",
                        "example": "No 5 Jesus Street"
                    },
                    "dob": {
                        "title": "dob",
                        "description": "date of birth",
                        "type": "string",
                        "example": "1994-49-33"
                    },
                    "notary_commission_number": {
                        "title": "Notary Commission Number",
                        "description": "Notary Commission Number of the notary",
                        "type": "string",
                        "example": "39403940349"
                    },
                    "country_id": {
                        "title": "Country Id",
                        "description": "Country id",
                        "type": "string",
                        "example": "39403940349"
                    },
                    "state_id": {
                        "title": "state Id",
                        "description": "state id",
                        "type": "string",
                        "example": "39403940349"
                    },
                    "city_id": {
                        "title": "city Id",
                        "description": "city id",
                        "type": "string",
                        "example": "39403940349"
                    },
                    "is_online": {
                        "title": "is_online",
                        "description": "is_online",
                        "type": "string",
                        "example": "true/false"
                    }
                },
                "type": "object"
            },
...
            "StoreDocumentResendOtpFormRequest": {
                "title": "User Document Resend OTP Form Request Fields",
                "description": "User Document Resend OTP request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "email": {
                        "title": "User reset email",
                        "description": "Reset email of the user",
                        "type": "string",
                        "example": "info@tonote.com"
                    },
                    "document_id": {
                        "title": "document_id",
                        "description": "document_id",
                        "type": "string",
                        "example": "Untitled"
                    }
                },
                "type": "object"
            },
...
            "UserLoginFormRequest": {
                "title": "Sign in Form Request Fields",
                "description": "sign in request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "user@tonote.com"
                    },
                    "password": {
                        "title": "User password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    },
                    "entry_point": {
                        "title": "User entry point",
                        "description": "entry point of the user",
                        "type": "string",
                        "example": "entry_point"
                    }
                },
                "type": "object"
            },
            "UserRegistrationFormRequest": {
                "title": "Sign Up Form Request Fields",
                "description": "sign up request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "description": "first_name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "last_name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "role": {
                        "title": "User Role",
                        "description": "User/Admin",
                        "type": "string",
                        "example": "User"
                    },
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "info@tonote.com"
                    },
                    "password": {
                        "title": "User password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}